+++ src/components/SIGV4ConnectionConfig.tsx
